refactor: Rename token list
