+++ listing.md
@@ -13,7 +13,7 @@ If you are seeking to be added to the [Pancakeswap.finance](https://pancakeswap.
 1. Fork the [PancakeSwap Interface](https://github.com/pancakeswap/pancake-swap-interface/) repository on GitHub
 2. Upload your **logo** in a 96\*96px transparent .png format to the `/public/images/coins` directory. Please name the logo as your symbol in uppercase.
 
-3. Add your **token information** using the example format provided below to `pancakeswap.json` in the `/src/constants/token/` directory.
+3. Add your **token information** using the example format provided below to `pantherswap.json` in the `/src/constants/token/` directory.
 
 4. Create a **pull request** detailing information about your project, website address, and contact details (telegram)
