+++ src/constants/token/pantherswap.json
@@ -1,5 +1,5 @@
 {
-  "name": "pancakeswap",
+  "name": "pantherswap",
   "timestamp": "2020-08-25T15:41:29.665Z",
   "version": {
     "major": 1,
