+++ src/state/lists/reducer.ts
@@ -5,7 +5,7 @@ import { TokenList } from '@uniswap/token-lists/dist/types'
 import { DEFAULT_LIST_OF_LISTS, DEFAULT_TOKEN_LIST_URL } from '../../constants/lists'
 import { updateVersion } from '../global/actions'
 import { acceptListUpdate, addList, fetchTokenList, removeList, selectList } from './actions'
-import DEFAULT_LIST from '../../constants/token/pancakeswap.json'
+import DEFAULT_LIST from '../../constants/token/pantherswap.json'
 
 export interface ListsState {
   readonly byUrl: {
