+++ src/utils/getTokenList.ts
@@ -7,7 +7,7 @@ import uriToHttp from './uriToHttp'
 
 // bakeryswap defaultTokenJson
 import { DEFAULT_TOKEN_LIST_URL } from '../constants/lists'
-import defaultTokenJson from '../constants/token/pancakeswap.json'
+import defaultTokenJson from '../constants/token/pantherswap.json'
 
 const tokenListValidator = new Ajv({ allErrors: true }).compile(schema)
